organized navigation types in a separate file
