+++ features/user/signin-view/signin-view.component.tsx
@@ -5,7 +5,7 @@ import {style} from '../../shared-style/style';
 import {userContext} from '../../context/market.context';
 import {useContext} from 'react';
 import type {NativeStackScreenProps} from '@react-navigation/native-stack';
-import type {RootStackParamList} from '../user-page/user-page.component';
+import type {RootStackParamList} from '../../navigation/types';
 export const SigninView = ({navigation}: NativeStackScreenProps<RootStackParamList, 'Signin'>) => {
   const contextData = useContext(userContext);
   const formData: Partial<User> | undefined = {
